+++ Features/ArtworkFeature/ArtworkFeature/Source/ArtworkViewController.swift
@@ -8,103 +8,6 @@
 import UIKit
 import ArtworkFeatureInterface
 
-class UploadViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .green
-    }
-}
-
-// MARK: - Builder
-
-
-
-
-
-public final class ArtworkViewBuilder: ArtworkViewBuildable {
-    let navigationController: UINavigationController
-    
-    public init(navigationController: UINavigationController) {
-        self.navigationController = navigationController
-        
-        // TODO: - replace dependency
-    }
-    
-    public func build(with listener: ArtworkListener?) -> ArtworkRouting {
-        let viewController = ArtworkViewController()
-        
-        let interactor = ArtworkInteractor()
-        
-        let router = ArtworkRouter(
-            navigationController: navigationController,
-            viewController: viewController
-        )
-        
-        interactor.presenter = viewController
-        interactor.listener = listener
-        interactor.router = router
-        
-        return router
-    }
-}
-
-// MARK: - Router
-
-final class ArtworkRouter: ArtworkRouting {
-    
-    let navigationController: UINavigationController
-    var viewController: UIViewController?
-    
-    init(
-        navigationController: UINavigationController,
-        viewController: UIViewController?
-    ) {
-        self.navigationController = navigationController
-        self.viewController = viewController
-    }
-    
-    // MARK: - Upload
-    
-    func pushToUpload() {
-        print(#function)
-        
-        let viewController = UploadViewController()
-        navigationController.pushViewController(viewController, animated: true)
-    }
-    
-}
-
-// MARK: - Interactor
-
-protocol ArtworkInteractorable: AnyObject {
-    var router: ArtworkRouting? { get set }
-    var listener: ArtworkListener? { get set }
-}
-
-final class ArtworkInteractor: ArtworkInteractorable,
-                               ArtworkViewPresentableListener
-{
-    weak var router: ArtworkRouting?
-    weak var listener: ArtworkListener?
-    weak var presenter: ArtworkViewPresentable?
-    
-    init() {}
-    
-    func fetch() { }
-    
-    func itemSelected(indexPath: IndexPath) {
-        print(#file, #function)
-        
-        router?.pushToUpload()
-    }
-}
-
-// MARK: - Presneter
-
-protocol ArtworkViewPresentable: AnyObject {
-    var listener: ArtworkViewPresentableListener? { get set }
-}
-
 protocol ArtworkViewPresentableListener: AnyObject {
     func itemSelected(indexPath: IndexPath)
 }
